chore: Build startup scripts instead of using tsx
For some reason, tsx could not exit gracefully when running these.
Either it got stuck, or exited but left the process running.
These issues did not occur with ts-node,
but the solution in this commit seemed more acceptable
than having a new dependency.
Future work could include investigating
if this is an issue with the server.